Added GPGSA Sentence and some tests. The integration is not completely done yet (Some times some tasks get aborted)

+++ src/faultInjection/fault_library/byte_manipulation/BytePerturbationFunction.java
@@ -5,11 +5,7 @@ public interface BytePerturbationFunction {
     String asString();
     long asLong();
 
-    default int allBitsLow() {
-        return ((int) 0);
-    }
+    default int allBitsLow() {return 0;}
 
-    default int allBitsHigh() {
-        return ((int) ~0);
-    }
+    default int allBitsHigh() {return  ~0;}
 }
